move socket boilerplate to boardpage
