+++ app/src/Home.tsx
@@ -1,12 +1,9 @@
-import { useEffect } from "react";
 import "./App.css";
 import "@radix-ui/themes/styles.css";
 import { Text } from "@radix-ui/themes";
-import { socket } from "./socket";
 import { PageContainer } from "./PageContainer";
 
 export function HomePage(): JSX.Element {
-  // const [isConnected, setIsConnected] = useState<boolean>(socket.connected);
   // const [isViewTransition, setIsViewTransition] = useState("");
 
   // const createRoom = ({})
@@ -17,24 +14,6 @@ export function HomePage(): JSX.Element {
   // if (document.startViewTransition) {
   //   setIsViewTransition("Yess, Your browser support View Transitions API");
   // }
-
-  useEffect(() => {
-    function onConnect(): void {
-      // setIsConnected(true);
-    }
-
-    function onDisconnect(): void {
-      // setIsConnected(false);
-    }
-
-    socket.on("connect", onConnect);
-    socket.on("disconnect", onDisconnect);
-
-    return () => {
-      socket.off("connect", onConnect);
-      socket.off("disconnect", onDisconnect);
-    };
-  }, []);
 
   return (
     <PageContainer title={"Welcome to Bike Bingo"}>
